Improve error handling and overpost protection

+++ application/Pitcher/Controllers/JobsController.cs
@@ -26,6 +26,7 @@ namespace Pitcher.Controllers
         }
 
         // GET: Jobs/Details/5
+        // COPY AND PASTE THIS METHOD CUSTOMIZATION INTO OTHER CONTROLLERS.  
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -34,6 +35,7 @@ namespace Pitcher.Controllers
             }
 
             var job = await _context.Jobs
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (job == null)
             {
@@ -52,15 +54,27 @@ namespace Pitcher.Controllers
         // POST: Jobs/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // COPY AND PASTE THIS METHOD CUSTOMIZATION INTO OTHER CONTROLLERS.  
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,JobTitle,JobDescription,JobStartDate,JobDeadline,JobIsComplete")] Job job)
+        public async Task<IActionResult> Create([Bind("JobTitle,JobDescription,JobStartDate,JobDeadline,JobIsComplete")] Job job)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(job);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(job);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                
+            }
+            catch(DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log).
+                ModelState.AddModelError("", "Unable to save changes. " +
+                "Try again, and if the problem persists " +
+                "See your system administrator.");
             }
             return View(job);
         }
@@ -84,40 +98,44 @@ namespace Pitcher.Controllers
         // POST: Jobs/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        // COPY AND PASTE THIS METHOD CUSTOMIZATION INTO OTHER CONTROLLERS. Prevents overposting.
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,JobTitle,JobDescription,JobStartDate,JobDeadline,JobIsComplete")] Job job)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != job.ID)
+            if (id == null)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            var jobToUpdate = await _context.Jobs.FirstOrDefaultAsync(j => j.ID == id);
+            if (await TryUpdateModelAsync<Job>(
+                jobToUpdate,
+                "",
+                j => j.JobTitle, j => j.JobDescription, j => j.JobStartDate, j => j.JobDeadline, j => j.JobIsComplete)) 
+            
             {
                 try
                 {
-                    _context.Update(job);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateConcurrencyException /* ex */)
                 {
-                    if (!JobExists(job.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (uncomment ex variable name and write a log.)
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                        "Try again, and if the problem persists, " +
+                        "see your system administrator.");
                 }
                 return RedirectToAction(nameof(Index));
             }
-            return View(job);
+            return View(jobToUpdate);
         }
 
         // GET: Jobs/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        // COPY AND PASTE THIS METHOD CUSTOMIZATION INTO OTHER CONTROLLERS.
+        // Catches delete failures and allows for retry.
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -125,24 +143,47 @@ namespace Pitcher.Controllers
             }
 
             var job = await _context.Jobs
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (job == null)
             {
                 return NotFound();
+            }
+            
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again, and if the problem persists " +
+                    "see your system administrator.";
             }
 
             return View(job);
         }
 
         // POST: Jobs/Delete/5
+        // COPY AND PASTE THIS METHOD CUSTOMIZATION INTO OTHER CONTROLLERS.
+        //Catches delete failures and allows for retry.
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var job = await _context.Jobs.FindAsync(id);
-            _context.Jobs.Remove(job);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if(job == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Jobs.Remove(job);
+                //SQL DELETE command is generated
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch(DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                return RedirectToAction(nameof(Delete), new {id = id, saveChangesError = true});
+            }
         }
 
         private bool JobExists(int id)
